1.1.3: Electron!!
Added electron to the script and made a test open. Everything works fine!

Right now the window doesn't do anything but in the next update there will be a live feed on trades ect.

bot.js: Changed the file to main.js

Webpage: Folder for holding all the assets required to run the user interface.

start.bat: Changed the starting code so it would work with electron

install.bat: Installs electron globally as well as the bot files.