1.2.1: App Styling & Sockets
Main.js can now communicate with the client and send a live update feed.

Index.html is now much better looking, it has some stats and a live update feed of succeded trades.
Planning on adding trades that were not accepted.

Removed folder "dev"- no further comments.

Added more styling in style.css for the index page.

Went back 1 update for socket.io in order to have the same version as client.

Added some more labels on the app. The "exit" button is not fully working and you should be able to exit your application with it.

Main.js sets up variable for the trade partners steam64 id and sends it to client.