1.2.2: Update game focus.
Added modal and working button that you can set your game that you want to base the bot around. There is now 4 game preset, I'm of course planning to add more.

Installed underscore.

Added the file configGames.html and set up the correct size of the window and options in order for it to work.

The game is updated in realtime which means that you can change it whenever you want.

BUG: The bot crashes when trading unwanted items that the bot can't look up online. I will have to work on this tomorrow.