Live update config :red_circle:
You can now live update your configs while the UI is running. This is thanks to a new module that I created that will read the values of the config file when the function is called.  :cherries:

Also fixed major bug that caused the config.json file to be reset. :bug: